Remove some warnings

+++ src/database/HttpRequestHandlers/ConsultRequestsByAreaHandler.hpp
@@ -25,9 +25,9 @@ class ConsultRequestsByAreaHandler : public DatabaseRequestHandler {
       // pack all data into a json format string
       std::stringstream requestsAsJSON;
       requestsAsJSON << "{";
-      for (int i = 0; i < requestsData.size(); ++i) {
+      for (size_t i = 0; i < requestsData.size(); ++i) {
         requestsAsJSON << "\"request" << (i + 1) << "\": { ";
-        for (int j = 0; j < requestsData[i].size(); ++j) {
+        for (size_t j = 0; j < requestsData[i].size(); ++j) {
           std::string jsonEntry;
           switch(j) {
             case 0:
